refactor: PlaceService를 TuripService로 통합 및 관련 의존성 수정
